requirements.txt
pymupdf>=1.24
Pillow>=10.0
